[FIX] move logic around
