+++ tests/conftest.py
@@ -37,15 +37,6 @@ from libpip2pi.commands import dir2pi
 
 pytest_plugins = ["pytester"]
 
-if typing.TYPE_CHECKING:
-    # Local type stub for mypy that works with both pytest < 7 and pytest >=7
-    # https://docs.pytest.org/en/7.1.x/_modules/_pytest/legacypath.html#TempdirFactory
-    import py
-
-    class TempdirFactory:
-        def mktemp(self, path: str) -> py.path.local:
-            ...
-
 
 @pytest.fixture(autouse=True)
 def set_cwd(testdir):
@@ -125,10 +116,3 @@ def examples_v2_package_index(pytestconfig) -> Iterator[str]:
             )
         dir2pi(argv=["dir2pi", artifact_dir])
         yield os.path.join(artifact_dir, "simple")
-
-
-@pytest.fixture(scope="function")
-def inmanta_state_dir(tmpdir_factory: "TempdirFactory") -> Iterator[str]:
-    inmanta_state_dir = tmpdir_factory.mktemp("inmanta_state_dir")
-    yield str(inmanta_state_dir)
-    inmanta_state_dir.remove(ignore_errors=True)
